Moved patching to own class - added patching for file integrity checks
