+++ Source/OpenCAGE/OpenCAGE.csproj
@@ -74,6 +74,7 @@
     <Compile Include="LaunchGame.designer.cs">
       <DependentUpon>LaunchGame.cs</DependentUpon>
     </Compile>
+    <Compile Include="PatchManager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs">
