+++ Source/OpenCAGE/Properties/AssemblyInfo.cs
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.9.2.3")]
-[assembly: AssemblyFileVersion("0.9.2.3")]
+[assembly: AssemblyVersion("0.9.2.31")]
+[assembly: AssemblyFileVersion("0.9.2.31")]
